dbft: fix timeout calculation in InitializeConsensus
Remove `blockPersistTime` field and calculate timeout value
based on current view only. Time passed between block persist and
calling `InitializeConsensus` is of the order of milliseconds and is
assumed to be negligible.
This is needed because blocks can be acquired via network.
In this case persisting is done outside of dbft library.