context: correctly save previous change view payloads during view reset
They're used to create recovery messages and currently we're completely
missing CV payloads there because of this logic. In fact we want to save
messages that lead to the current view we're resetting the state for (==), but
newer ones don't hurt also (>=), even though we shouldn't have them. See C#
implementation.