dbft: add the notion of pending change view
When we reject CV because we have data about other nodes they do similar thing
and if there is no proposal they all do this thing which means they wait for
another timer event to actually send CV. This makes the system delay for more
than expected (30s+60s instead of 30s for 15s blocks). Pending view change
fixes this by "remembering" that we actually can send a CV if there are enough
alive nodes on the network.