models: adjust MoreThanFNodesCommitted check
Based on the https://github.com/neo-project/neo-modules/issues/792#issuecomment-1455759793,
the definition of MoreThanFNodesCommitted should be adjusted to match
the core algorithm as it's the key factor of the four-good-nodes liveness lock
scenario. There following change is made:

* We consider the node to be "Committed" if it has the Commit message
  sent at _any_ view. If the good node has committed, then we know for sure
  that it won't go further to the next consensus round and can rely on
  this information.

The thing that remains the same is that we do not count the "lost" nodes, because
this information can't be reliably trusted. See the comment inside the commit.

Based on this adjustment, the first liveness lock scenario mentioned in
https://github.com/neo-project/neo-modules/issues/792#issue-1609058923
("Liveness lock with four non-faulty nodes") needs to include one more step
to enter a deadlock: one of the replicas that in the "cv" state must die in the
end of the scenario.

Moreover, there's another liveness lock scenario when one of the nodes is in
the RMDead list, i.e. can "die" at any moment. Consider running the base model
specification with the following configuration:
```
RM      	RMFault 	RMDead 	MaxView
{0, 1, 2, 3} 	{}      	{0} 	2
```