Revert "timer: do not drain the timer channel"
This reverts commit f7a23ed23f2dbbd21dceb7c1de64611965eb3127.

Draining is needed because there's a possibility that timer is requested to be reset with `0` delta:
https://github.com/nspcc-dev/dbft/blob/5281a9ae584e866e6dfa2a99313f7db6818c8e9d/dbft.go#L157-L159

Consider the situation when CV happens:  the timer will be reset in ChangeView message handler:
https://github.com/nspcc-dev/dbft/blob/5281a9ae584e866e6dfa2a99313f7db6818c8e9d/check.go#L179

Which means that dBFT eventloop is still busy with CV handling and can't read from the timer's channel (given that `d` is `0`):
https://github.com/nspcc-dev/dbft/blob/5281a9ae584e866e6dfa2a99313f7db6818c8e9d/timer/timer.go#L60-L65

And this channel has a buffer of `1` in case if it's internal Timer-bound
channel, which is not enough to finish CV processing. So we end up in a deadlock.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>